Share service level funcitons in example CLI to reduce duplicated code
