+++ src/iovation.LaunchKey.Sdk.ExampleCli/OrgExamples.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Threading;
 using iovation.LaunchKey.Sdk.Error;
 
 namespace iovation.LaunchKey.Sdk.ExampleCli
@@ -9,37 +8,7 @@ namespace iovation.LaunchKey.Sdk.ExampleCli
         public static int DoServiceAuth(string orgId, string privateKey, string serviceId, string userId, string apiURL)
         {
             var serviceClient = ClientFactories.MakeOrganizationServiceClient(orgId, privateKey, serviceId, apiURL);
-            try
-            {
-                var authorizationRequest = serviceClient.CreateAuthorizationRequest(userId);
-                while (true)
-                {
-                    Console.WriteLine("checking auth");
-
-                    // poll for a response
-                    var authResponse = serviceClient.GetAuthorizationResponse(authorizationRequest.Id);
-
-                    // if we got one, process it
-                    if (authResponse != null)
-                    {
-                        Console.WriteLine($"Auth response was {authResponse.Authorized}");
-                        return 0;
-                    }
-
-                    // if not, we are still waiting on the user. wait a bit ... 
-                    Thread.Sleep(1000);
-                }
-            }
-            catch (AuthorizationRequestTimedOutError)
-            {
-                Console.WriteLine("user never replied.");
-                return 1;
-            }
-            catch (BaseException e)
-            {
-                Console.WriteLine($"Error while authorizing user {userId} against service ID {serviceId}. Error: {e.Message}");
-                return 1;
-            }
+            return SharedServiceHelpers.DoAuthorizationRequest(serviceClient, userId);
         }
 
         public static int DoDirectoryDeviceList(string orgId, string privateKey, string directoryId, string userId, string apiURL)
