+++ src/iovation.LaunchKey.Sdk.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27703.2042
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "iovation.LaunchKey.Sdk.ExampleCli", "iovation.LaunchKey.Sdk.ExampleCli\iovation.LaunchKey.Sdk.ExampleCli.csproj", "{98609471-5035-4A52-893C-8A156FC8751B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "iovation.LaunchKey.Sdk.ExampleCli", "iovation.LaunchKey.Sdk.ExampleCli\iovation.LaunchKey.Sdk.ExampleCli.csproj", "{98609471-5035-4A52-893C-8A156FC8751B}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{0297C3A8-E07B-45D8-9C3E-24C3A2FFFE1E}"
 	ProjectSection(SolutionItems) = preProject
